+++ lib/markdown/toolbar/modal_input_url.dart
@@ -66,7 +66,7 @@ class ModalInputUrl extends StatelessWidget {
                           color: Colors.white,
                         ),
                       ),
-                      backgroundColor: Colors.red.withOpacity(0.8),
+                      backgroundColor: Colors.red.withValues(alpha: 0.8),
                       duration: const Duration(milliseconds: 700),
                     ),
                   );