+++ lib/ui/components/dialogs/libraries_dialog.dart
@@ -34,7 +34,7 @@ void showLibrariesDialog(BuildContext context) {
                         color: Theme.of(context)
                             .colorScheme
                             .onSurface
-                            .withOpacity(0.5)),
+                            .withValues(alpha: 0.5)),
                   ),
                 ),
               );