+++ lib/ui/screens/archive/archive_screen.dart
@@ -104,7 +104,10 @@ class _ArchiveScreenState extends State<ArchiveScreen> {
             size: 48,
             color: isDirectory
                 ? Theme.of(context).colorScheme.secondary
-                : Theme.of(context).colorScheme.secondary.withOpacity(0.8),
+                : Theme.of(context)
+                    .colorScheme
+                    .secondary
+                    .withValues(alpha: 0.8),
           ),
           title: Text(
             name,