+++ lib/ui/screens/notes/note_editor.dart
@@ -324,7 +324,7 @@ class _NoteEditorScreenState extends State<NoteEditorScreen> {
           color: Theme.of(context).colorScheme.surface,
           boxShadow: [
             BoxShadow(
-                color: Colors.grey.withOpacity(0.2),
+                color: Colors.grey.withValues(alpha: 0.2),
                 spreadRadius: 5,
                 blurRadius: 7,
                 offset: const Offset(0, 3)),